Move Value, SimpleValue and SimpleType to serde
