+++ serde_dhall/src/serde.rs
@@ -5,9 +5,9 @@ use serde::de::value::{
 };
 
 use dhall::syntax::NumKind;
-use dhall::{SValKind, SimpleValue};
 
-use crate::{Deserialize, Error, Result};
+use crate::simple::{SValKind, SimpleValue};
+use crate::{Deserialize, Error, Result, Value};
 
 impl<'a, T> crate::sealed::Sealed for T where T: serde::Deserialize<'a> {}
 
@@ -17,7 +17,7 @@ impl<'a, T> Deserialize for T
 where
     T: serde::Deserialize<'a>,
 {
-    fn from_dhall(v: &dhall::Value) -> Result<Self> {
+    fn from_dhall(v: &Value) -> Result<Self> {
         let sval = v.to_simple_value().ok_or_else(|| {
             Error::Deserialize(format!(
                 "this cannot be deserialized into the serde data model: {}",
